ACL - ARO and ACO along with Access component

+++ Controller/PositionsController.php
@@ -18,37 +18,69 @@ class PositionsController extends AppController
 
     function view($ID = NULL)
     {
-        $this->set('position', $this->Position->read(NULL, $ID));
+        $user = $this->Auth->user();
+        if(!$this->Acl->check($user['role'], 'Position', 'view'))
+            //if(!$this->Access->check('User', 'view'))
+        {
+            die('You are not authorized ');
+        }
+        else {
+            $this->set('position', $this->Position->read(NULL, $ID));
+        }
     }
 
     function add()
     {
-        if (!empty($this->data)) {
-            if ($this->Position->save($this->data)) {
-                $this->Session->setFlash('Position sucessfuly added');
-                $this->redirect(array('position' => 'index'));
-            } else {
-                $this->Session->setFlash('Position unsucessfuly added');
+        $user = $this->Auth->user();
+        if(!$this->Acl->check($user['role'], 'Position', 'create'))
+            //if(!$this->Access->check('User', 'view'))
+        {
+            die('You are not authorized ');
+        }
+        else {
+            if (!empty($this->data)) {
+                if ($this->Position->save($this->data)) {
+                    $this->Session->setFlash('Position sucessfuly added');
+                    $this->redirect(array('position' => 'index'));
+                } else {
+                    $this->Session->setFlash('Position unsucessfuly added');
+                }
             }
         }
     }
 
     function edit($ID = NULL)
     {
-        if (empty($this->data)) {
-            $this->data = $this->Position->read(NULL, $ID);
-        } else {
-            if ($this->Position->save($this->data)) {
-                $this->Session->setFlash('Position updated');
-                $this->redirect(array('action' => 'view', $ID));
+        $user = $this->Auth->user();
+        if(!$this->Acl->check($user['role'], 'Position', 'update'))
+            //if(!$this->Access->check('User', 'view'))
+        {
+            die('You are not authorized ');
+        }
+        else {
+            if (empty($this->data)) {
+                $this->data = $this->Position->read(NULL, $ID);
+            } else {
+                if ($this->Position->save($this->data)) {
+                    $this->Session->setFlash('Position updated');
+                    $this->redirect(array('action' => 'view', $ID));
+                }
             }
         }
     }
 
     function delete($ID = NULL)
     {
-        $this->Position->delete($ID);
-        $this->Session->setFlash('Position deleted');
-        $this->redirect(array('action' => 'index'));
+        $user = $this->Auth->user();
+        if(!$this->Acl->check($user['role'], 'Position', 'delete'))
+            //if(!$this->Access->check('User', 'view'))
+        {
+            die('You are not authorized ');
+        }
+        else {
+            $this->Position->delete($ID);
+            $this->Session->setFlash('Position deleted');
+            $this->redirect(array('action' => 'index'));
+        }
     }
 }
